change codeword_size from int to size_t

+++ node.c
@@ -21,7 +21,7 @@ void remove_from_linked_list(node *delete_node){
 	
 }
 
-node *traverse_n_nodes_forward(node *current_node, int n) {
+node *traverse_n_nodes_forward(node *current_node, size_t n) {
 	int index;
 	for (index = 0; index<n; index++) {
 		current_node = current_node->next;
@@ -29,7 +29,7 @@ node *traverse_n_nodes_forward(node *current_node, int n) {
 	return current_node;
 }
 
-node *traverse_n_nodes_backward(node *current_node, int n) {
+node *traverse_n_nodes_backward(node *current_node, size_t n) {
 	int index;
 	for (index = 0; index<n; index++) {
 		current_node = current_node->prev;
